+++ NEWS.md
@@ -6,6 +6,7 @@
 * Update title and description in package DESCRIPTION file (#139)
 * Add `Median_Quartiles` to `pairwise_test_cont()` output (#99)
 * Improve documentation for datasets and fix typos in various documentation files (#148)
+* Add geometric mean, median, quantile, and standard deviation functions (#82)
 * Minor Changes:
   * Update `Overview()` to include new `Median_Quartiles` (#99)
 
@@ -14,7 +15,6 @@
 * Report loaded and attached packages in package reproducibility table (#132)
 * Report nodename in platform reproducibility table (#121)
 * Add B cell example dataset from G001 flow and sequencing file (#108)
-* Add geometric mean, median, quantile, and standard deviation functions (#82)
 * Maintenance of package and CI (#114, #116, #124)
 * VISCfunctions now depends on R >= 4.1.0 (#134)
 * Auto-generate VISCfunctions pkgdown website with GitHub Actions (#119)