+++ NEWS.md
@@ -8,6 +8,7 @@
 * Improve documentation for datasets and fix typos in various documentation files (#148)
 * Minor Changes:
   * Update `Overview()` to include new `Median_Quartiles` (#99)
+* Fix hardcoded floating point issues causing unexpected behavior in `round_away_0()` (#156)
 
 # VISCfunctions 1.3.1