+++ NEWS.md
@@ -1,4 +1,4 @@
-# VISCfunctions (development version)
+# VISCfunctions 1.3.2
 
 * Add PR template for the package (#138)
 * Add `SampleSizes` output to `pairwise_test_bin()` (#130)